requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Unipath==1.1
PyYAML==6.0
